pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "This is end to end RAG project."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
   "ensure>=1.0.4",
   "fastapi>=0.115.12",
   "ipykernel>=6.29.5",
   "jinja2>=3.1.6",
   "langchain>=0.3.24",
   "langchain-core>=0.3.56",
   "langchain-groq>=0.3.2",
   "langchain-huggingface>=0.1.2",
   "langchain-pinecone>=0.2.5",
   "pandas>=2.2.3",
   "python-box>=7.3.2",
   "python-dotenv>=1.1.0",
   "python-multipart>=0.0.20",
   "pyyaml>=6.0.2",
   "uvicorn>=0.34.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
